+++ src/components/Rating/Rating.css
@@ -1,0 +1,4 @@
+.starRating {
+    display: flex;
+    justify-content: center;
+}
