+++ src/components/RatingList/RatingList.js
@@ -2,69 +2,9 @@ import React from 'react';
 import Rating from '../Rating/Rating'
 import RatingForm from '../RatingForm/RatingForm';
 import { Route, Switch, Link } from 'react-router-dom';
-// import font awesome
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-// solid Star
-import { faStar as solidStar } from '@fortawesome/free-solid-svg-icons';
-// holy Star
-import { faStar as holyStar } from '@fortawesome/free-regular-svg-icons';
 
 function RatingList(props) {
-	// handles Star Ratings
-	function star1(favStar) {
-		if (favStar === 0) {
-			return (
-				<div>
-					<FontAwesomeIcon id='1' icon={holyStar} size='2x' />
-				</div>
-			);
-		} else {
-			return (
-				<div>
-					<FontAwesomeIcon id='1' icon={solidStar} size='2x' />
-				</div>
-			);
-		}
-	}
-	function star2(favStar) {
-		if (favStar > 1) {
-			return (
-				<div>
-					<FontAwesomeIcon id='2' icon={solidStar} size='2x' />
-				</div>
-			);
-		} else {
-			return (
-				<div>
-					<FontAwesomeIcon id='2' icon={holyStar} size='2x' />
-				</div>
-			);
-		}
-	}
-	function star3(favStar) {
-		if (favStar > 2) {
-			return (
-				<div>
-					<FontAwesomeIcon id='3' icon={solidStar} size='2x' />
-				</div>
-			);
-		} else {
-			return (
-				<div>
-					<FontAwesomeIcon id='3' icon={holyStar} size='2x' />
-				</div>
-			);
-		}
-	}
-	function stars(rating) {
-		return (
-			<>
-				{star1(rating)}
-				{star2(rating)}
-				{star3(rating)}
-			</>
-		);
-	}
+	
 	// will need to check pathing
 	const id = props.match.params.id;
